feat: implement identification scanner for passport. refactor: sharedStyles
